refactor: 拆分 dashboard.js
