+++ assets/dashboard.js
@@ -1,2499 +1,1 @@
-        // ECharts instance variables
-        let chartTraffic = null;
-        let chartBandwidth = null;
-        let chartRequests = null;
-        let chartPerformance = null;
-        let chartSecurity = null;
-        let chartOriginPull = null;
-        let chartTopCountry = null;
-        let chartTopProvince = null;
-        let chartTopStatusCode = null;
-        let chartTopDomain = null;
-        let chartTopUrl = null;
-        let chartTopResourceType = null;
-        let chartTopSip = null;
-        let chartTopReferer = null;
-        let chartTopUaDevice = null;
-        let chartTopUaBrowser = null;
-        let chartTopUaOs = null;
-        let chartTopUa = null;
-        let chartTopRequestCountry = null;
-        let chartTopRequestProvince = null;
-        let chartTopRequestStatusCode = null;
-        let chartTopRequestDomain = null;
-        let chartTopRequestUrl = null;
-        let chartTopRequestResourceType = null;
-        let chartTopRequestSip = null;
-        let chartTopRequestReferer = null;
-        let chartTopRequestUaDevice = null;
-        let chartTopRequestUaBrowser = null;
-        let chartTopRequestUaOs = null;
-        let chartTopRequestUa = null;
-        let chartTopMap = null;
-        let chartFunctionRequests = null;
-        let chartFunctionCpu = null;
-
-        const provinceMap = {
-            '22': '北京', '86': '内蒙古', '146': '山西', '1069': '河北', '1177': '天津',
-            '119': '宁夏', '152': '陕西', '1208': '甘肃', '1467': '青海', '1468': '新疆',
-            '145': '黑龙江', '1445': '吉林', '1464': '辽宁', '2': '福建', '120': '江苏',
-            '121': '安徽', '122': '山东', '1050': '上海', '1442': '浙江', '182': '河南',
-            '1135': '湖北', '1465': '江西', '1466': '湖南', '118': '贵州', '153': '云南',
-            '1051': '重庆', '1068': '四川', '1155': '西藏', '4': '广东', '173': '广西',
-            '1441': '海南', '0': '其他', '1': '港澳台', '-1': '境外'
-        };
-
-        const countryMap = {
-            'CN': '中国大陆', 'AF': '阿富汗', 'MV': '马尔代夫', 'AM': '亚美尼亚', 'MN': '蒙古', 'AZ': '阿塞拜疆',
-            'MM': '缅甸', 'BH': '巴林', 'NP': '尼泊尔', 'BD': '孟加拉', 'KP': '朝鲜',
-            'BT': '不丹', 'OM': '阿曼', 'IO': '英属印度洋领地', 'PK': '巴基斯坦', 'KH': '柬埔寨',
-            'PS': '巴勒斯坦', 'CX': '圣诞岛', 'PH': '菲律宾', 'HK': '中国香港', 'QA': '卡塔尔',
-            'IN': '印度', 'SA': '沙特阿拉伯', 'ID': '印度尼西亚', 'SG': '新加坡', 'IR': '伊朗',
-            'KR': '韩国', 'IQ': '伊拉克', 'LK': '斯里兰卡', 'IL': '以色列', 'SY': '叙利亚',
-            'JP': '日本', 'TW': '中国台湾', 'JO': '约旦', 'TJ': '塔吉克斯坦', 'KZ': '哈萨克斯坦',
-            'TH': '泰国', 'KW': '科威特', 'TM': '土库曼斯坦', 'KG': '吉尔吉斯斯坦', 'AE': '阿联酋',
-            'LA': '老挝', 'UZ': '乌兹别克斯坦', 'LB': '黎巴嫩', 'VN': '越南', 'MO': '中国澳门',
-            'YE': '也门', 'MY': '马来西亚', 'TR': '土耳其', 'AX': '奥兰群岛', 'IT': '意大利',
-            'AL': '阿尔巴尼亚', 'JE': '泽西岛', 'AD': '安道尔', 'LT': '立陶宛', 'AT': '奥地利',
-            'LU': '卢森堡', 'BY': '白俄罗斯', 'MK': '马其顿', 'BE': '比利时', 'MT': '马耳他',
-            'BA': '波黑', 'MD': '摩尔多瓦', 'BG': '保加利亚', 'MC': '摩纳哥', 'BQ': '荷兰加勒比区',
-            'ME': '黑山', 'HR': '克罗地亚', 'NL': '荷兰', 'CZ': '捷克', 'NO': '挪威',
-            'DK': '丹麦', 'PL': '波兰', 'EE': '爱沙尼亚', 'PT': '葡萄牙', 'FO': '法罗群岛',
-            'RO': '罗马尼亚', 'FI': '芬兰', 'RU': '俄罗斯', 'FR': '法国', 'SM': '圣马力诺',
-            'DE': '德国', 'RS': '塞尔维亚', 'GI': '直布罗陀', 'SX': '荷属圣马丁', 'GR': '希腊',
-            'SK': '斯洛伐克', 'GG': '根西岛', 'ES': '西班牙', 'HU': '匈牙利', 'SE': '瑞典',
-            'IS': '冰岛', 'CH': '瑞士', 'IE': '爱尔兰', 'UA': '乌克兰', 'IM': '马恩岛',
-            'GB': '英国', 'DZ': '阿尔及利亚', 'ML': '马里', 'AO': '安哥拉', 'MR': '毛里塔尼亚',
-            'BJ': '贝宁', 'MU': '毛里求斯', 'BW': '博茨瓦纳', 'YT': '马约特', 'BF': '布基纳法索',
-            'MA': '摩洛哥', 'BI': '布隆迪', 'MZ': '莫桑比克', 'CM': '喀麦隆', 'NA': '纳米比亚',
-            'CV': '佛得角', 'NE': '尼日尔', 'CF': '中非', 'NG': '尼日利亚', 'TD': '乍得',
-            'RW': '卢旺达', 'KM': '科摩罗', 'SH': '圣赫勒拿', 'DJ': '吉布提', 'ST': '圣多美和普林西比',
-            'EG': '埃及', 'SN': '塞内加尔', 'GQ': '赤道几内亚', 'SC': '塞舌尔', 'ER': '厄立特里亚',
-            'SL': '塞拉利昂', 'ET': '埃塞俄比亚', 'SO': '索马里', 'GA': '加蓬', 'ZA': '南非',
-            'GM': '冈比亚', 'SS': '南苏丹', 'GH': '加纳', 'SD': '苏丹', 'GN': '几内亚',
-            'SZ': '斯威士兰', 'GW': '几内亚比绍', 'TZ': '坦桑尼亚', 'KE': '肯尼亚', 'TG': '多哥',
-            'LS': '莱索托', 'TN': '突尼斯', 'LR': '利比里亚', 'UG': '乌干达', 'LY': '利比亚',
-            'EH': '西撒哈拉', 'MG': '马达加斯加', 'ZM': '赞比亚', 'MW': '马拉维', 'ZW': '津巴布韦',
-            'CD': '刚果民主共和国', 'CG': '刚果共和国', 'CI': '科特迪瓦', 'AU': '澳大利亚', 'NF': '诺福克岛',
-            'CK': '库克群岛', 'MP': '北马里亚纳群岛', 'TL': '东帝汶', 'PW': '帕劳', 'GU': '关岛',
-            'PG': '巴布亚新几内亚', 'KI': '基里巴斯', 'SB': '所罗门群岛', 'MH': '马绍尔群岛', 'TO': '汤加',
-            'NR': '瑙鲁', 'TV': '图瓦卢', 'NZ': '新西兰', 'AI': '安圭拉', 'HT': '海地',
-            'AG': '安提瓜和巴布达', 'HN': '洪都拉斯', 'AW': '阿鲁巴', 'JM': '牙买加', 'BS': '巴哈马',
-            'MX': '墨西哥', 'BB': '巴巴多斯', 'MS': '蒙塞拉特岛', 'BM': '百慕大', 'NI': '尼加拉瓜',
-            'CA': '加拿大', 'PA': '巴拿马', 'KY': '开曼群岛', 'PR': '波多黎各', 'CR': '哥斯达黎加',
-            'KN': '圣基茨和尼维斯', 'CU': '古巴', 'LC': '圣卢西亚', 'CW': '库拉索', 'MF': '法属圣马丁',
-            'SV': '萨尔瓦多', 'TT': '特立尼达和多巴哥', 'GL': '格陵兰岛', 'TC': '特克斯和凯科斯群岛',
-            'GD': '格林纳达', 'US': '美国', 'GT': '危地马拉', 'AR': '阿根廷', 'GY': '圭亚那',
-            'BO': '玻利维亚', 'PY': '巴拉圭', 'BR': '巴西', 'PE': '秘鲁', 'CL': '智利',
-            'SR': '苏里南', 'CO': '哥伦比亚', 'UY': '乌拉圭', 'EC': '厄瓜多尔', 'VE': '委内瑞拉',
-            'GF': '法属圭亚那', 'Antarctica': '南极洲'
-        };
-
-        const codeToMapName = {
-            'CN': 'China', 'AF': 'Afghanistan', 'AL': 'Albania', 'DZ': 'Algeria', 'AO': 'Angola', 'AR': 'Argentina',
-            'AM': 'Armenia', 'AU': 'Australia', 'AT': 'Austria', 'AZ': 'Azerbaijan', 'BS': 'Bahamas', 'BH': 'Bahrain',
-            'BD': 'Bangladesh', 'BB': 'Barbados', 'BY': 'Belarus', 'BE': 'Belgium', 'BZ': 'Belize', 'BJ': 'Benin',
-            'BT': 'Bhutan', 'BO': 'Bolivia', 'BA': 'Bosnia and Herzegovina', 'BW': 'Botswana', 'BR': 'Brazil',
-            'BN': 'Brunei', 'BG': 'Bulgaria', 'BF': 'Burkina Faso', 'BI': 'Burundi', 'KH': 'Cambodia', 'CM': 'Cameroon',
-            'CA': 'Canada', 'CF': 'Central African Republic', 'TD': 'Chad', 'CL': 'Chile', 'CO': 'Colombia',
-            'KM': 'Comoros', 'CG': 'Republic of the Congo', 'CD': 'Democratic Republic of the Congo', 'CR': 'Costa Rica',
-            'HR': 'Croatia', 'CU': 'Cuba', 'CY': 'Cyprus', 'CZ': 'Czech Republic', 'DK': 'Denmark', 'DJ': 'Djibouti',
-            'DO': 'Dominican Republic', 'EC': 'Ecuador', 'EG': 'Egypt', 'SV': 'El Salvador', 'GQ': 'Equatorial Guinea',
-            'ER': 'Eritrea', 'EE': 'Estonia', 'ET': 'Ethiopia', 'FJ': 'Fiji', 'FI': 'Finland', 'FR': 'France',
-            'GA': 'Gabon', 'GM': 'Gambia', 'GE': 'Georgia', 'DE': 'Germany', 'GH': 'Ghana', 'GR': 'Greece',
-            'GT': 'Guatemala', 'GN': 'Guinea', 'GW': 'Guinea-Bissau', 'GY': 'Guyana', 'HT': 'Haiti', 'HN': 'Honduras',
-            'HU': 'Hungary', 'IS': 'Iceland', 'IN': 'India', 'ID': 'Indonesia', 'IR': 'Iran', 'IQ': 'Iraq',
-            'IE': 'Ireland', 'IL': 'Israel', 'IT': 'Italy', 'CI': 'Ivory Coast', 'JM': 'Jamaica', 'JP': 'Japan',
-            'JO': 'Jordan', 'KZ': 'Kazakhstan', 'KE': 'Kenya', 'KP': 'North Korea', 'KR': 'South Korea',
-            'KW': 'Kuwait', 'KG': 'Kyrgyzstan', 'LA': 'Laos', 'LV': 'Latvia', 'LB': 'Lebanon', 'LS': 'Lesotho',
-            'LR': 'Liberia', 'LY': 'Libya', 'LT': 'Lithuania', 'LU': 'Luxembourg', 'MK': 'Macedonia', 'MG': 'Madagascar',
-            'MW': 'Malawi', 'MY': 'Malaysia', 'ML': 'Mali', 'MT': 'Malta', 'MR': 'Mauritania', 'MU': 'Mauritius',
-            'MX': 'Mexico', 'MD': 'Moldova', 'MN': 'Mongolia', 'ME': 'Montenegro', 'MA': 'Morocco', 'MZ': 'Mozambique',
-            'MM': 'Myanmar', 'NA': 'Namibia', 'NP': 'Nepal', 'NL': 'Netherlands', 'NZ': 'New Zealand', 'NI': 'Nicaragua',
-            'NE': 'Niger', 'NG': 'Nigeria', 'NO': 'Norway', 'OM': 'Oman', 'PK': 'Pakistan', 'PA': 'Panama',
-            'PG': 'Papua New Guinea', 'PY': 'Paraguay', 'PE': 'Peru', 'PH': 'Philippines', 'PL': 'Poland',
-            'PT': 'Portugal', 'PR': 'Puerto Rico', 'QA': 'Qatar', 'RO': 'Romania', 'RU': 'Russia', 'RW': 'Rwanda',
-            'SA': 'Saudi Arabia', 'SN': 'Senegal', 'RS': 'Serbia', 'SL': 'Sierra Leone', 'SG': 'Singapore',
-            'SK': 'Slovakia', 'SI': 'Slovenia', 'SB': 'Solomon Islands', 'SO': 'Somalia', 'ZA': 'South Africa',
-            'SS': 'South Sudan', 'ES': 'Spain', 'LK': 'Sri Lanka', 'SD': 'Sudan', 'SR': 'Suriname', 'SZ': 'Swaziland',
-            'SE': 'Sweden', 'CH': 'Switzerland', 'SY': 'Syria', 'TW': 'Taiwan', 'TJ': 'Tajikistan', 'TZ': 'Tanzania',
-            'TH': 'Thailand', 'TL': 'Timor-Leste', 'TG': 'Togo', 'TT': 'Trinidad and Tobago', 'TN': 'Tunisia',
-            'TR': 'Turkey', 'TM': 'Turkmenistan', 'UG': 'Uganda', 'UA': 'Ukraine', 'AE': 'United Arab Emirates',
-            'GB': 'United Kingdom', 'US': 'United States', 'UY': 'Uruguay', 'UZ': 'Uzbekistan', 'VU': 'Vanuatu',
-            'VE': 'Venezuela', 'VN': 'Vietnam', 'YE': 'Yemen', 'ZM': 'Zambia', 'ZW': 'Zimbabwe'
-        };
-
-        const worldNameMap = {};
-        Object.keys(codeToMapName).forEach(code => {
-            if (countryMap[code]) {
-                worldNameMap[codeToMapName[code]] = countryMap[code];
-            }
-        });
-
-        const metricsConfig = {
-            traffic: ['l7Flow_flux', 'l7Flow_inFlux', 'l7Flow_outFlux'],
-            bandwidth: ['l7Flow_bandwidth', 'l7Flow_inBandwidth', 'l7Flow_outBandwidth'],
-            originPull: ['l7Flow_outFlux_hy', 'l7Flow_inFlux_hy', 'l7Flow_outBandwidth_hy', 'l7Flow_inBandwidth_hy', 'l7Flow_request_hy'],
-            requests: ['l7Flow_request'],
-            performance: ['l7Flow_avgResponseTime', 'l7Flow_avgFirstByteResponseTime'],
-            edgeFunctions: ['function_requestCount', 'function_cpuCostTime'],
-            security: ['ccAcl_interceptNum', 'ccManage_interceptNum', 'ccRate_interceptNum'],
-            topAnalysis: ['l7Flow_outFlux_country', 'l7Flow_outFlux_province', 'l7Flow_outFlux_statusCode', 'l7Flow_outFlux_domain', 'l7Flow_outFlux_url', 'l7Flow_outFlux_resourceType', 'l7Flow_outFlux_sip', 'l7Flow_outFlux_referers', 'l7Flow_outFlux_ua_device', 'l7Flow_outFlux_ua_browser', 'l7Flow_outFlux_ua_os', 'l7Flow_outFlux_ua', 'l7Flow_request_country', 'l7Flow_request_province', 'l7Flow_request_statusCode', 'l7Flow_request_domain', 'l7Flow_request_url', 'l7Flow_request_resourceType', 'l7Flow_request_sip', 'l7Flow_request_referers', 'l7Flow_request_ua_device', 'l7Flow_request_ua_browser', 'l7Flow_request_ua_os', 'l7Flow_request_ua']
-        };
-
-        const metricLabels = {
-            'l7Flow_flux': '总流量',
-            'l7Flow_inFlux': '客户端请求流量',
-            'l7Flow_outFlux': '响应流量',
-            'l7Flow_bandwidth': '总带宽',
-            'l7Flow_inBandwidth': '请求带宽',
-            'l7Flow_outBandwidth': '响应带宽',
-            'l7Flow_outFlux_hy': '回源请求流量',
-            'l7Flow_inFlux_hy': '回源响应流量',
-            'l7Flow_outBandwidth_hy': '回源请求带宽',
-            'l7Flow_inBandwidth_hy': '回源响应带宽',
-            'l7Flow_request_hy': '回源请求数',
-            'l7Flow_request': '请求数',
-            'l7Flow_avgResponseTime': '平均响应耗时',
-            'l7Flow_avgFirstByteResponseTime': '平均首字节耗时',
-            'function_requestCount': 'Edge Functions 请求数',
-            'function_cpuCostTime': 'Edge Functions CPU 时间',
-            'ccAcl_interceptNum': '精确防护拦截',
-            'ccManage_interceptNum': '托管规则拦截',
-            'ccRate_interceptNum': '速率限制拦截',
-            'l7Flow_outFlux_country': '国家/地区流量',
-            'l7Flow_outFlux_province': '国内省份流量',
-            'l7Flow_outFlux_statusCode': '状态码流量',
-            'l7Flow_outFlux_domain': '域名流量',
-            'l7Flow_outFlux_url': 'URL 流量',
-            'l7Flow_outFlux_resourceType': '资源类型流量',
-            'l7Flow_outFlux_sip': '客户端IP流量',
-            'l7Flow_outFlux_referers': 'Referer 流量',
-            'l7Flow_outFlux_ua_device': '设备类型流量',
-            'l7Flow_outFlux_ua_browser': '浏览器流量',
-            'l7Flow_outFlux_ua_os': '操作系统流量',
-            'l7Flow_outFlux_ua': 'User Agent 流量',
-            'l7Flow_request_country': '国家/地区请求数',
-            'l7Flow_request_province': '国内省份请求数',
-            'l7Flow_request_statusCode': '状态码请求数',
-            'l7Flow_request_domain': '域名请求数',
-            'l7Flow_request_url': 'URL 请求数',
-            'l7Flow_request_resourceType': '资源类型请求数',
-            'l7Flow_request_sip': '客户端IP请求数',
-            'l7Flow_request_referers': 'Referer 请求数',
-            'l7Flow_request_ua_device': '设备类型请求数',
-            'l7Flow_request_ua_browser': '浏览器请求数',
-            'l7Flow_request_ua_os': '操作系统请求数',
-            'l7Flow_request_ua': 'User Agent 请求数'
-        };
-
-        const metricColors = {
-            'l7Flow_flux': '#3b82f6', // blue
-            'l7Flow_inFlux': '#f59e0b', // amber
-            'l7Flow_outFlux': '#10b981', // green
-            'l7Flow_bandwidth': '#8b5cf6', // purple
-            'l7Flow_inBandwidth': '#ec4899', // pink
-            'l7Flow_outBandwidth': '#06b6d4', // cyan
-            'l7Flow_outFlux_hy': '#3b82f6', // blue
-            'l7Flow_inFlux_hy': '#10b981', // green
-            'l7Flow_outBandwidth_hy': '#8b5cf6', // purple
-            'l7Flow_inBandwidth_hy': '#ec4899', // pink
-            'l7Flow_request_hy': '#f43f5e', // rose
-            'l7Flow_request': '#f43f5e', // rose
-            'l7Flow_avgResponseTime': '#ef4444', // red
-            'l7Flow_avgFirstByteResponseTime': '#f97316', // orange
-            'function_requestCount': '#8b5cf6', // purple
-            'function_cpuCostTime': '#06b6d4', // cyan
-            'ccAcl_interceptNum': '#ef4444', // red
-            'ccManage_interceptNum': '#f59e0b', // amber
-            'ccRate_interceptNum': '#3b82f6', // blue
-            'l7Flow_outFlux_country': '#3b82f6', // blue
-            'l7Flow_outFlux_province': '#f59e0b', // amber
-            'l7Flow_outFlux_statusCode': '#8b5cf6', // purple
-            'l7Flow_outFlux_domain': '#06b6d4', // cyan
-            'l7Flow_outFlux_url': '#10b981', // green
-            'l7Flow_outFlux_resourceType': '#f59e0b', // amber
-            'l7Flow_outFlux_sip': '#ef4444', // red
-            'l7Flow_outFlux_referers': '#8b5cf6', // purple
-            'l7Flow_outFlux_ua_device': '#06b6d4', // cyan
-            'l7Flow_outFlux_ua_browser': '#f59e0b', // amber
-            'l7Flow_outFlux_ua_os': '#10b981', // green
-            'l7Flow_outFlux_ua': '#8b5cf6', // purple
-            'l7Flow_request_country': '#3b82f6', // blue
-            'l7Flow_request_province': '#f59e0b', // amber
-            'l7Flow_request_statusCode': '#8b5cf6', // purple
-            'l7Flow_request_domain': '#06b6d4', // cyan
-            'l7Flow_request_url': '#10b981', // green
-            'l7Flow_request_resourceType': '#f59e0b', // amber
-            'l7Flow_request_sip': '#8b5cf6', // purple
-            'l7Flow_request_referers': '#ec4899', // pink
-            'l7Flow_request_ua_device': '#06b6d4', // cyan
-            'l7Flow_request_ua_browser': '#10b981', // green
-            'l7Flow_request_ua_os': '#f59e0b', // amber
-            'l7Flow_request_ua': '#8b5cf6' // purple
-        };
-
-        /** Tool Functions **/
-
-        function formatDate(date) {
-            return date.toISOString().slice(0, 19) + 'Z';
-        }
-
-        function handleTimeRangeChange() {
-            const range = document.getElementById('timeRange').value;
-            const customInputs = document.getElementById('customTimeInputs');
-            if (range === 'custom') {
-                customInputs.classList.remove('hidden');
-                customInputs.classList.add('flex');
-            } else {
-                customInputs.classList.add('hidden');
-                customInputs.classList.remove('flex');
-                refreshData();
-            }
-        }
-
-        function calculateTimeRange() {
-            const range = document.getElementById('timeRange').value;
-            const now = new Date();
-            let endTime = new Date(now);
-            let startTime;
-
-            switch(range) {
-                case '30min': startTime = new Date(now.getTime() - 30 * 60 * 1000); break;
-                case '1h': startTime = new Date(now.getTime() - 1 * 60 * 60 * 1000); break;
-                case '6h': startTime = new Date(now.getTime() - 6 * 60 * 60 * 1000); break;
-                case 'today': 
-                    startTime = new Date(now);
-                    startTime.setHours(0, 0, 0, 0);
-                    break;
-                case 'yesterday':
-                    startTime = new Date(now);
-                    startTime.setDate(now.getDate() - 1);
-                    startTime.setHours(0, 0, 0, 0);
-                    
-                    endTime = new Date(now);
-                    endTime.setDate(now.getDate() - 1);
-                    endTime.setHours(23, 59, 59, 999);
-                    break;
-                case '3d': startTime = new Date(now.getTime() - 3 * 24 * 60 * 60 * 1000); break;
-                case '7d': startTime = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000); break;
-                case '14d': startTime = new Date(now.getTime() - 14 * 24 * 60 * 60 * 1000); break;
-                case '31d': startTime = new Date(now.getTime() - 31 * 24 * 60 * 60 * 1000); break;
-                case 'custom':
-                    const days = parseInt(document.getElementById('customDays').value) || 0;
-                    const hours = parseInt(document.getElementById('customHours').value) || 0;
-                    const minutes = parseInt(document.getElementById('customMinutes').value) || 0;
-                    const seconds = parseInt(document.getElementById('customSeconds').value) || 0;
-                    
-                    let totalMs = ((days * 24 * 60 * 60) + (hours * 60 * 60) + (minutes * 60) + seconds) * 1000;
-                    const maxMs = 31 * 24 * 60 * 60 * 1000; // 31 days limit
-                    const errorEl = document.getElementById('timeRangeError');
-
-                    if (totalMs > maxMs) {
-                        if (errorEl) errorEl.innerText = '自定义时间范围不能超过 31 天，已自动为您调整为 31 天。';
-                        totalMs = maxMs;
-                    } else {
-                        if (errorEl) errorEl.innerText = '';
-                    }
-
-                    if (totalMs > 0) {
-                        startTime = new Date(now.getTime() - totalMs);
-                    } else {
-                        // Default to 1 hour if nothing entered
-                        startTime = new Date(now.getTime() - 60 * 60 * 1000);
-                    }
-                    break;
-                default: startTime = new Date(now.getTime() - 24 * 60 * 60 * 1000);
-            }
-
-            return {
-                startTime: formatDate(startTime),
-                endTime: formatDate(endTime)
-            };
-        }
-
-        async function fetchData(metric) {
-            try {
-                const { startTime, endTime } = calculateTimeRange();
-                const interval = document.getElementById('interval').value;
-                const zoneId = document.getElementById('zoneId').value.trim();
-                
-                let url = `/api/traffic?metric=${metric}&startTime=${encodeURIComponent(startTime)}&endTime=${encodeURIComponent(endTime)}`;
-                if (interval && interval !== 'auto') {
-                    url += `&interval=${interval}`;
-                }
-                if (zoneId) {
-                    url += `&zoneId=${encodeURIComponent(zoneId)}`;
-                }
-
-                const response = await fetch(url);
-                const result = await response.json();
-                if (result.error) throw new Error(result.error);
-                return result;
-            } catch (err) {
-                console.error(`Error fetching ${metric}:`, err);
-                return null;
-            }
-        }
-
-        function processData(result, targetMetric) {
-            // Check for Top Data Structure (DescribeTopL7AnalysisData)
-            // It returns Data[0].DetailData (Array of Key/Value)
-            if (result?.Data?.[0]?.DetailData) {
-                 return {
-                     type: 'top',
-                     data: result.Data[0].DetailData
-                 };
-            }
-
-            // DescribeTimingL7AnalysisData returns 'Data'
-            // DescribeTimingL7OriginPullData returns 'TimingDataRecords'
-            const dataList = result?.Data || result?.TimingDataRecords || [];
-            if (dataList.length === 0) return { timeData: [], valueData: [], sum: 0, max: 0, avg: 0, type: 'time' };
-
-            let typeValue;
-            
-            // Try to find matching metric in TypeValue or Value
-            if (targetMetric) {
-                if (dataList[0]?.TypeValue) {
-                    typeValue = dataList[0].TypeValue.find(item => item.MetricName === targetMetric);
-                } else if (dataList[0]?.Value) {
-                    typeValue = dataList[0].Value.find(item => item.MetricName === targetMetric);
-                }
-            }
-
-            // Fallback if not found or no targetMetric (or if only one exists)
-            if (!typeValue) {
-                 typeValue = dataList[0]?.TypeValue?.[0];
-                 // Support DescribeWebProtectionData structure (Value instead of TypeValue)
-                 if (!typeValue && dataList[0]?.Value?.[0]) {
-                    typeValue = dataList[0].Value[0];
-                 }
-            }
-
-            const details = typeValue?.Detail || [];
-            const sum = typeValue?.Sum || 0;
-            const max = typeValue?.Max || 0;
-            const avg = typeValue?.Avg || 0;
-
-            const timeData = [];
-            const valueData = [];
-
-            // Context for formatting
-            const range = document.getElementById('timeRange')?.value || '30min';
-            const interval = document.getElementById('interval')?.value || 'auto';
-            const longRanges = ['3d', '7d', '14d', '31d'];
-            
-            // Auto-detect if interval is effectively 'day'
-            let isDayInterval = interval === 'day';
-            if (interval === 'auto' && details.length > 1) {
-                const diff = details[1].Timestamp - details[0].Timestamp;
-                // If interval is >= 23 hours (allow some slack), treat as day
-                if (diff >= 82800) isDayInterval = true;
-            }
-            // If auto and range is 31d/15d, it usually defaults to day
-            if (interval === 'auto' && ['14d', '31d'].includes(range)) {
-                // Check if we have few points (e.g. < 32 for 31d), likely day
-                if (details.length <= 32) isDayInterval = true;
-            }
-
-            // Calculate time span of the data
-            let span = 0;
-            if (details.length > 0) {
-                 span = details[details.length - 1].Timestamp - details[0].Timestamp;
-            }
-
-            details.forEach(item => {
-                const date = new Date(item.Timestamp * 1000);
-                let label;
-
-                if (isDayInterval) {
-                    // Show YYYY-MM-DD
-                    label = `${date.getFullYear()}-${(date.getMonth()+1).toString().padStart(2, '0')}-${date.getDate().toString().padStart(2, '0')}`;
-                } else if (span > 86400 || longRanges.includes(range) || (range === 'custom' && span > 86400)) {
-                    // Show MM-DD HH:mm for long ranges or custom > 24h
-                    label = `${(date.getMonth()+1).toString().padStart(2, '0')}-${date.getDate().toString().padStart(2, '0')} ${date.getHours().toString().padStart(2, '0')}:${date.getMinutes().toString().padStart(2, '0')}`;
-                } else {
-                    // Show HH:mm for short ranges
-                    label = `${date.getHours().toString().padStart(2, '0')}:${date.getMinutes().toString().padStart(2, '0')}`;
-                }
-
-                timeData.push(label);
-                valueData.push(item.Value);
-            });
-
-            return { timeData, valueData, sum, max, avg, type: 'time' };
-        }
-
-        function formatBytes(bytes) {
-            if (bytes === 0) return '0 B';
-            const k = 1024;
-            const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-            const i = Math.floor(Math.log(bytes) / Math.log(k));
-            if (i < 0) return bytes + ' B';
-            if (i >= sizes.length) return (bytes / Math.pow(k, sizes.length - 1)).toFixed(2) + ' ' + sizes[sizes.length - 1];
-            return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
-        }
-
-        function formatBps(bps) {
-            if (bps === 0) return '0 bps';
-            const k = 1024;
-            const sizes = ['bps', 'Kbps', 'Mbps', 'Gbps', 'Tbps', 'Pbps', 'Ebps', 'Zbps', 'Ybps'];
-            const i = Math.floor(Math.log(bps) / Math.log(k));
-            if (i < 0) return bps + ' bps';
-            if (i >= sizes.length) return (bps / Math.pow(k, sizes.length - 1)).toFixed(2) + ' ' + sizes[sizes.length - 1];
-            return parseFloat((bps / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
-        }
-
-        function formatCount(num) {
-            if (num === 0) return '0';
-            if (num < 1000) return num.toString();
-            if (num < 10000) return (num / 1000).toFixed(2) + ' 千';
-            if (num < 100000000) return (num / 10000).toFixed(2) + ' 万';
-            return (num / 100000000).toFixed(2) + ' 亿';
-        }
-
-        function getBestCountUnit(maxValue) {
-            if (maxValue < 1000) return { unit: '次', divisor: 1 };
-            if (maxValue < 10000) return { unit: '千次', divisor: 1000 };
-            if (maxValue < 100000000) return { unit: '万次', divisor: 10000 };
-            return { unit: '亿次', divisor: 100000000 };
-        }
-
-        function getBestUnit(maxValue, type = 'bytes') {
-            const k = 1024;
-            const byteUnits = ['B', 'KB', 'MB', 'GB', 'TB', 'PB'];
-            const bitUnits = ['bps', 'Kbps', 'Mbps', 'Gbps', 'Tbps', 'Pbps'];
-            const units = type === 'bandwidth' ? bitUnits : byteUnits;
-
-            if (maxValue === 0) return { unit: units[0], divisor: 1 };
-
-            let i = Math.floor(Math.log(maxValue) / Math.log(k));
-            if (i < 0) i = 0;
-            if (i >= units.length) i = units.length - 1;
-
-            return { unit: units[i], divisor: Math.pow(k, i) };
-        }
-
-        /** Feature Functions **/
-
-        async function fetchZones() {
-            try {
-                const response = await fetch('/api/zones');
-                const result = await response.json();
-                
-                const select = document.getElementById('zoneId');
-                // Keep the current selection if possible, or default to *
-                const currentVal = select.value;
-                select.innerHTML = ''; 
-
-                // Always add "All Zones" option
-                const allOption = document.createElement('option');
-                allOption.value = "*";
-                allOption.text = "全部站点";
-                select.appendChild(allOption);
-
-                if (result.Zones && result.Zones.length > 0) {
-                    result.Zones.forEach(zone => {
-                        const option = document.createElement('option');
-                        option.value = zone.ZoneId;
-                        let text = zone.ZoneName;
-                        if (text === 'default-pages-zone') {
-                            text += ' (Pages站点)';
-                        }
-                        option.text = text;
-                        select.appendChild(option);
-                    });
-                }
-                
-                // Restore selection if it exists in new options, otherwise default to *
-                if (currentVal && Array.from(select.options).some(o => o.value === currentVal)) {
-                    select.value = currentVal;
-                } else {
-                    select.value = "*";
-                }
-
-            } catch (err) {
-                console.error("Error fetching zones:", err);
-                const select = document.getElementById('zoneId');
-                // If error, ensure we have at least the default * option
-                if (select.options.length === 0 || select.options[0].value !== '*') {
-                     select.innerHTML = '<option value="*">获取站点失败 (默认为全部)</option>';
-                }
-            }
-        }
-
-        ////ENTRANCE////
-        async function initDashboard() {
-            // Initialize charts
-            chartTraffic = echarts.init(document.getElementById('chart_traffic'));
-            chartBandwidth = echarts.init(document.getElementById('chart_bandwidth'));
-            chartRequests = echarts.init(document.getElementById('chart_requests'));
-            chartPerformance = echarts.init(document.getElementById('chart_performance'));
-            chartSecurity = echarts.init(document.getElementById('chart_security'));
-            chartOriginPull = echarts.init(document.getElementById('chart_origin_pull'));
-            chartTopCountry = echarts.init(document.getElementById('chart_top_country'));
-            chartTopProvince = echarts.init(document.getElementById('chart_top_province'));
-            chartTopStatusCode = echarts.init(document.getElementById('chart_top_status_code'));
-            chartTopDomain = echarts.init(document.getElementById('chart_top_domain'));
-            chartTopUrl = echarts.init(document.getElementById('chart_top_url'));
-            chartTopResourceType = echarts.init(document.getElementById('chart_top_resource_type'));
-            chartTopSip = echarts.init(document.getElementById('chart_top_sip'));
-            chartTopReferer = echarts.init(document.getElementById('chart_top_referer'));
-            chartTopUaDevice = echarts.init(document.getElementById('chart_top_ua_device'));
-            chartTopUaBrowser = echarts.init(document.getElementById('chart_top_ua_browser'));
-            chartTopUaOs = echarts.init(document.getElementById('chart_top_ua_os'));
-            chartTopUa = echarts.init(document.getElementById('chart_top_ua'));
-            chartTopRequestCountry = echarts.init(document.getElementById('chart_top_request_country'));
-            chartTopRequestProvince = echarts.init(document.getElementById('chart_top_request_province'));
-            chartTopRequestStatusCode = echarts.init(document.getElementById('chart_top_request_status_code'));
-            chartTopRequestDomain = echarts.init(document.getElementById('chart_top_request_domain'));
-            chartTopRequestUrl = echarts.init(document.getElementById('chart_top_request_url'));
-            chartTopRequestResourceType = echarts.init(document.getElementById('chart_top_request_resource_type'));
-            chartTopRequestSip = echarts.init(document.getElementById('chart_top_request_sip'));
-            chartTopRequestReferer = echarts.init(document.getElementById('chart_top_request_referer'));
-            chartTopRequestUaDevice = echarts.init(document.getElementById('chart_top_request_ua_device'));
-            chartTopRequestUaBrowser = echarts.init(document.getElementById('chart_top_request_ua_browser'));
-            chartTopRequestUaOs = echarts.init(document.getElementById('chart_top_request_ua_os'));
-            chartTopRequestUa = echarts.init(document.getElementById('chart_top_request_ua'));
-            chartTopMap = echarts.init(document.getElementById('chart_top_map'));
-            chartFunctionRequests = echarts.init(document.getElementById('chart_function_requests'));
-            chartFunctionCpu = echarts.init(document.getElementById('chart_function_cpu'));
-            chartPagesCloudFunctionRequests = echarts.init(document.getElementById('chart_pages_cloud_function_requests'));
-
-            await fetchZones(); // Fetch zones first
-            await refreshData();
-        }
-
-        async function fetchPagesBuildStats() {
-             try {
-                // Pages build stats are global, not per zone (usually), or the API handles it.
-                // If it needs zoneId, we might need to pass it.
-                // The current backend implementation just calls DescribePagesResources without params, 
-                // so it likely returns account-level or default zone stats.
-                // However, the user request implied it's a new API. 
-                // Let's assume it doesn't need time range or zoneId for now as per backend implementation.
-                
-                // Get ZoneId if selected
-                const zoneIdElement = document.getElementById('zoneId');
-                const zoneId = zoneIdElement ? zoneIdElement.value : null;
-                const url = zoneId && zoneId !== '*' ? `/api/pages/build-count?zoneId=${zoneId}` : '/api/pages/build-count';
-
-                const response = await fetch(url);
-                const result = await response.json();
-                
-                if (result.parsedResult) {
-                    const { dplDailyCount, dplMonthCount } = result.parsedResult;
-                    document.getElementById('kpi_pages_daily_build').innerText = dplDailyCount !== undefined ? dplDailyCount : '-';
-                    document.getElementById('kpi_pages_monthly_build').innerText = dplMonthCount !== undefined ? dplMonthCount : '-';
-                } else {
-                     document.getElementById('kpi_pages_daily_build').innerText = '-';
-                     document.getElementById('kpi_pages_monthly_build').innerText = '-';
-                }
-
-            } catch (err) {
-                console.error("Error fetching pages build stats:", err);
-                document.getElementById('kpi_pages_daily_build').innerText = 'Error';
-                document.getElementById('kpi_pages_monthly_build').innerText = 'Error';
-            }
-        }
-
-        async function fetchPagesCloudFunctionStats(startTime, endTime) {
-            try {
-                // Get ZoneId if selected
-                const zoneIdElement = document.getElementById('zoneId');
-                const zoneId = zoneIdElement ? zoneIdElement.value : null;
-                
-                let url = zoneId && zoneId !== '*' ? `/api/pages/cloud-function-requests?zoneId=${zoneId}` : '/api/pages/cloud-function-requests';
-                
-                // Add time range params if provided
-                if (startTime && endTime) {
-                    const separator = url.includes('?') ? '&' : '?';
-                    url += `${separator}startTime=${encodeURIComponent(startTime)}&endTime=${encodeURIComponent(endTime)}`;
-                }
-
-                const response = await fetch(url);
-                const result = await response.json();
-                
-                // Parsed Result: { Status, Granularity, Timestamps, Values, TotalValue }
-                if (result.parsedResult) {
-                    const { TotalValue, Timestamps, Values } = result.parsedResult;
-                    
-                    // Update KPI
-                    document.getElementById('kpi_pages_cloud_function_total').innerText = TotalValue !== undefined ? TotalValue : '-';
-                    
-                    // Update Chart
-                    if (Timestamps && Values && Timestamps.length > 0) {
-                         const dates = Timestamps.map(ts => {
-                            // Timestamps are in seconds, convert to local string
-                            return new Date(ts * 1000).toLocaleString('zh-CN', { month: 'numeric', day: 'numeric', hour: 'numeric', minute: 'numeric' });
-                        });
-
-                        const option = {
-                            tooltip: {
-                                trigger: 'axis'
-                            },
-                            grid: {
-                                left: '3%',
-                                right: '4%',
-                                bottom: '3%',
-                                containLabel: true
-                            },
-                            xAxis: {
-                                type: 'category',
-                                boundaryGap: false,
-                                data: dates
-                            },
-                            yAxis: {
-                                type: 'value'
-                            },
-                            series: [
-                                {
-                                    name: 'Requests',
-                                    type: 'line',
-                                    smooth: true,
-                                    data: Values,
-                                    areaStyle: {
-                                        opacity: 0.1
-                                    },
-                                    itemStyle: {
-                                        color: '#3b82f6'
-                                    }
-                                }
-                            ]
-                        };
-                        chartPagesCloudFunctionRequests.setOption(option);
-                    } else {
-                        chartPagesCloudFunctionRequests.clear();
-                    }
-
-                } else {
-                     document.getElementById('kpi_pages_cloud_function_total').innerText = '-';
-                     chartPagesCloudFunctionRequests.clear();
-                }
-            } catch (err) {
-                console.error("Error fetching pages cloud function stats:", err);
-                document.getElementById('kpi_pages_cloud_function_total').innerText = 'Error';
-                chartPagesCloudFunctionRequests.clear();
-            }
-        }
-
-        async function fetchPagesCloudFunctionMonthlyStats() {
-            try {
-                // Get ZoneId if selected
-                const zoneIdElement = document.getElementById('zoneId');
-                const zoneId = zoneIdElement ? zoneIdElement.value : null;
-                const url = zoneId && zoneId !== '*' ? `/api/pages/cloud-function-monthly-stats?zoneId=${zoneId}` : '/api/pages/cloud-function-monthly-stats';
-                
-                const response = await fetch(url);
-                const result = await response.json();
-                
-                // Expected Result: { parsedResult: { TotalMemDuration: number, TotalInvocation: number } }
-                if (result.parsedResult) {
-                    const { TotalMemDuration, TotalInvocation } = result.parsedResult;
-                    
-                    // Update Requests KPI
-                    document.getElementById('kpi_pages_monthly_cf_requests').innerText = TotalInvocation !== undefined ? TotalInvocation : '-';
-                    
-                    // Update GBs KPI (TotalMemDuration / 1024)
-                    if (TotalMemDuration !== undefined) {
-                        const gbs = (TotalMemDuration / 1024).toFixed(2);
-                        document.getElementById('kpi_pages_monthly_cf_gbs').innerText = gbs;
-                    } else {
-                        document.getElementById('kpi_pages_monthly_cf_gbs').innerText = '-';
-                    }
-
-                } else {
-                     document.getElementById('kpi_pages_monthly_cf_requests').innerText = '-';
-                     document.getElementById('kpi_pages_monthly_cf_gbs').innerText = '-';
-                }
-            } catch (err) {
-                console.error("Error fetching pages cloud function monthly stats:", err);
-                document.getElementById('kpi_pages_monthly_cf_requests').innerText = 'Error';
-                document.getElementById('kpi_pages_monthly_cf_gbs').innerText = 'Error';
-            }
-        }
-
-        async function refreshData() {
-            // Show loading
-            document.querySelectorAll('[id^="kpi_"]').forEach(el => el.innerText = '加载中...');
-
-            // Check time range for Security Metrics
-            const { startTime, endTime } = calculateTimeRange();
-            const start = new Date(startTime);
-            const end = new Date(endTime);
-            // Allow a small buffer (e.g., 1 minute) for "7 days" check to handle slight offsets
-            const isSecuritySupported = (end - start) <= (14 * 24 * 60 * 60 * 1000 + 60000);
-
-            // Fetch all metrics
-            const allMetrics = [
-                ...metricsConfig.traffic,
-                ...metricsConfig.bandwidth,
-                ...metricsConfig.originPull,
-                ...metricsConfig.requests,
-                ...metricsConfig.performance,
-                ...metricsConfig.edgeFunctions,
-                ...(isSecuritySupported ? metricsConfig.security : []),
-                ...metricsConfig.topAnalysis
-            ];
-
-            const results = {};
-            
-            // Start Pages Build Stats fetch (independent)
-            fetchPagesBuildStats();
-            // Start Pages Cloud Function Stats fetch (independent) - Pass time range
-            fetchPagesCloudFunctionStats(startTime, endTime);
-            // Start Pages Cloud Function Monthly Stats fetch (independent)
-            fetchPagesCloudFunctionMonthlyStats();
-
-            // Parallel fetch
-            await Promise.all(allMetrics.map(async (metric) => {
-                const res = await fetchData(metric);
-                if (res) {
-                    results[metric] = processData(res, metric);
-                }
-            }));
-
-            // Update UI
-            updateTrafficSection(results); //1.
-            updateBandwidthSection(results); //2.
-            updateOriginPullSection(results); //3.
-            updateRequestsSection(results); //4.
-            updatePerformanceSection(results); //5.
-            updateEdgeFunctionsSection(results); // New
-            updateSecuritySection(results); //5.1
-            updateTopAnalysisSection(results); //6.
-        }
-
-        //1.有使用
-        function updateTrafficSection(results) {
-            const metrics = metricsConfig.traffic;
-            const series = [];
-            let timeData = [];
-
-            // Calculate global max
-            let globalMax = 0;
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (data && data.type === 'time') {
-                    const max = Math.max(...data.valueData);
-                    if (max > globalMax) globalMax = max;
-                }
-            });
-
-            const { unit, divisor } = getBestUnit(globalMax, 'bytes');
-
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (!data || data.type !== 'time') return;
-                
-                // Update KPI
-                document.getElementById(`kpi_${metric}`).innerText = formatBytes(data.sum);
-
-                if (timeData.length === 0) timeData = data.timeData;
-
-                // Chart Data (Dynamic Unit)
-                const valueData = data.valueData.map(v => (v / divisor).toFixed(2));
-                
-                series.push({
-                    name: metricLabels[metric],
-                    type: 'line',
-                    smooth: true,
-                    data: valueData,
-                    itemStyle: { color: metricColors[metric] },
-                    areaStyle: { opacity: 0.1 },
-                    // Store raw data for tooltip
-                    rawData: data.valueData
-                });
-            });
-
-            const option = {
-                tooltip: { trigger: 'axis', formatter: (params) => {
-                    let res = params[0].axisValue + '<br/>';
-                    params.forEach(param => {
-                        // Access raw data using dataIndex
-                        const seriesIndex = param.seriesIndex;
-                        const dataIndex = param.dataIndex;
-                        const rawVal = series[seriesIndex].rawData[dataIndex];
-                        const formattedVal = formatBytes(rawVal);
-                        
-                        res += `${param.marker}${param.seriesName}: ${formattedVal}<br/>`;
-                    });
-                    return res;
-                }},
-                legend: { data: metrics.map(m => metricLabels[m]), bottom: 0 },
-                grid: { left: '3%', right: '4%', bottom: '10%', top: '15%', containLabel: true },
-                xAxis: { type: 'category', boundaryGap: false, data: timeData },
-                yAxis: { type: 'value', name: unit },
-                series: series
-            };
-            chartTraffic.setOption(option);
-        }
-
-        //2.有使用
-        function updateBandwidthSection(results) {
-            const metrics = metricsConfig.bandwidth;
-            const series = [];
-            let timeData = [];
-
-            // Calculate global max
-            let globalMax = 0;
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (data && data.type === 'time') {
-                    const max = Math.max(...data.valueData);
-                    if (max > globalMax) globalMax = max;
-                }
-            });
-
-            const { unit, divisor } = getBestUnit(globalMax, 'bandwidth');
-
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (!data || data.type !== 'time') return;
-                
-                // Update KPI (Max/Peak)
-                document.getElementById(`kpi_${metric}`).innerText = formatBps(data.max);
-
-                if (timeData.length === 0) timeData = data.timeData;
-
-                // Chart Data (Dynamic Unit)
-                const valueData = data.valueData.map(v => (v / divisor).toFixed(2));
-                
-                series.push({
-                    name: metricLabels[metric],
-                    type: 'line',
-                    smooth: true,
-                    data: valueData,
-                    itemStyle: { color: metricColors[metric] },
-                    areaStyle: { opacity: 0.1 },
-                    // Store raw data
-                    rawData: data.valueData
-                });
-            });
-
-            const option = {
-                tooltip: { trigger: 'axis', formatter: (params) => {
-                    let res = params[0].axisValue + '<br/>';
-                    params.forEach(param => {
-                        const seriesIndex = param.seriesIndex;
-                        const dataIndex = param.dataIndex;
-                        const rawVal = series[seriesIndex].rawData[dataIndex];
-                        const formattedVal = formatBps(rawVal);
-                        res += `${param.marker}${param.seriesName}: ${formattedVal}<br/>`;
-                    });
-                    return res;
-                }},
-                legend: { data: metrics.map(m => metricLabels[m]), bottom: 0 },
-                grid: { left: '3%', right: '4%', bottom: '10%', top: '15%', containLabel: true },
-                xAxis: { type: 'category', boundaryGap: false, data: timeData },
-                yAxis: { type: 'value', name: unit },
-                series: series
-            };
-            chartBandwidth.setOption(option);
-        }
-
-        //3.有使用
-        function updateOriginPullSection(results) {
-            const metrics = metricsConfig.originPull;
-            const series = [];
-            let timeData = [];
-
-            // Calculate global max for Flux, Bandwidth and Requests separately
-            let maxFlux = 0;
-            let maxBandwidth = 0;
-            let maxRequests = 0;
-
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (data && data.type === 'time') {
-                    const max = Math.max(...data.valueData);
-                    if (metric.includes('Flux')) {
-                        if (max > maxFlux) maxFlux = max;
-                    } else if (metric.includes('Bandwidth')) {
-                        if (max > maxBandwidth) maxBandwidth = max;
-                    } else if (metric.includes('request')) {
-                        if (max > maxRequests) maxRequests = max;
-                    }
-                }
-            });
-
-            const fluxUnit = getBestUnit(maxFlux, 'bytes');
-            const bandwidthUnit = getBestUnit(maxBandwidth, 'bandwidth');
-            const requestUnit = getBestCountUnit(maxRequests);
-
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (!data || data.type !== 'time') return;
-                
-                // Update KPI
-                const kpiEl = document.getElementById(`kpi_${metric}`);
-                if (kpiEl) {
-                    if (metric.includes('Flux')) {
-                         kpiEl.innerText = formatBytes(data.sum);
-                    } else if (metric.includes('Bandwidth')) {
-                         kpiEl.innerText = formatBps(data.max);
-                    } else if (metric.includes('request')) {
-                         kpiEl.innerText = formatCount(data.sum);
-                    } else {
-                         kpiEl.innerText = data.sum.toLocaleString();
-                    }
-                }
-
-                if (timeData.length === 0) timeData = data.timeData;
-
-                let chartData = [];
-                let unit = '';
-                let divisor = 1;
-
-                // Normalize data for chart
-                if (metric.includes('Flux')) {
-                    unit = fluxUnit.unit;
-                    divisor = fluxUnit.divisor;
-                } else if (metric.includes('Bandwidth')) {
-                    unit = bandwidthUnit.unit;
-                    divisor = bandwidthUnit.divisor;
-                } else if (metric.includes('request')) {
-                    unit = requestUnit.unit;
-                    divisor = requestUnit.divisor;
-                } else {
-                    unit = '';
-                    divisor = 1;
-                }
-
-                chartData = data.valueData.map(v => {
-                    const val = v / divisor;
-                    return (divisor === 1) ? val : val.toFixed(2);
-                });
-
-                series.push({
-                    name: metricLabels[metric],
-                    type: 'line',
-                    smooth: true,
-                    data: chartData,
-                    itemStyle: { color: metricColors[metric] },
-                    areaStyle: { opacity: 0.1 },
-                    customUnit: unit,
-                    rawData: data.valueData
-                });
-            });
-
-            const option = {
-                title: { show: false },
-                tooltip: { 
-                    trigger: 'axis',
-                    formatter: (params) => {
-                        let res = params[0].axisValue + '<br/>';
-                        params.forEach(param => {
-                            const seriesIndex = param.seriesIndex;
-                            const dataIndex = param.dataIndex;
-                            const rawVal = series[seriesIndex].rawData[dataIndex];
-                            
-                            const unit = series[seriesIndex].customUnit || '';
-                            let formattedVal = '';
-                            
-                            if (unit.includes('bps')) {
-                                formattedVal = formatBps(rawVal);
-                            } else if (unit.includes('B')) {
-                                formattedVal = formatBytes(rawVal);
-                            } else if (unit.includes('千') || unit.includes('万') || unit.includes('亿')) {
-                                formattedVal = formatCount(rawVal);
-                            } else {
-                                formattedVal = rawVal.toLocaleString();
-                            }
-
-                            res += `${param.marker}${param.seriesName}: ${formattedVal}<br/>`;
-                        });
-                        return res;
-                    }
-                },
-                legend: { data: metrics.map(m => metricLabels[m]), bottom: 0 },
-                grid: { left: '3%', right: '4%', bottom: '10%', top: '15%', containLabel: true },
-                xAxis: { type: 'category', boundaryGap: false, data: timeData },
-                yAxis: { type: 'value' },
-                series: series
-            };
-            chartOriginPull.setOption(option);
-
-            // Calculate Cache Hit Rate
-            // Formula: 1 - (Origin Response Flux / EdgeOne Response Flux)
-            const originFluxData = results['l7Flow_inFlux_hy']; // Origin Response
-            const edgeFluxData = results['l7Flow_outFlux']; // EdgeOne Response
-
-            let hitRate = 0;
-            if (originFluxData && edgeFluxData && edgeFluxData.sum > 0) {
-                hitRate = 1 - (originFluxData.sum / edgeFluxData.sum);
-            }
-            // Clamp hitRate if necessary (e.g. if origin > edge due to compression diffs, it might be negative, but let's show real value or clamp?)
-            // Usually hit rate is 0-1.
-            
-            const hitRateEl = document.getElementById('kpi_cache_hit_rate');
-            if (hitRateEl) {
-                hitRateEl.innerText = (hitRate * 100).toFixed(2) + '%';
-            }
-        }
-
-        //4.有使用
-        function updateRequestsSection(results) {
-            const metrics = metricsConfig.requests;
-            const series = [];
-            let timeData = [];
-
-            // Calculate global max
-            let globalMax = 0;
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (data && data.type === 'time') {
-                    const max = Math.max(...data.valueData);
-                    if (max > globalMax) globalMax = max;
-                }
-            });
-
-            const { unit, divisor } = getBestCountUnit(globalMax);
-
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (!data || data.type !== 'time') return;
-                
-                // Update KPI
-                document.getElementById(`kpi_${metric}`).innerText = formatCount(data.sum);
-
-                if (timeData.length === 0) timeData = data.timeData;
-
-                // Chart Data (Dynamic Unit)
-                const valueData = data.valueData.map(v => {
-                    const val = v / divisor;
-                    return (divisor === 1) ? val : val.toFixed(2);
-                });
-
-                series.push({
-                    name: metricLabels[metric],
-                    type: 'line',
-                    smooth: true,
-                    data: valueData,
-                    itemStyle: { color: metricColors[metric] },
-                    areaStyle: { opacity: 0.2 },
-                    rawData: data.valueData
-                });
-            });
-
-            const option = {
-                tooltip: { trigger: 'axis', formatter: (params) => {
-                    let res = params[0].axisValue + '<br/>';
-                    params.forEach(param => {
-                        const seriesIndex = param.seriesIndex;
-                        const dataIndex = param.dataIndex;
-                        const rawVal = series[seriesIndex].rawData[dataIndex];
-                        const formattedVal = formatCount(rawVal);
-                        res += `${param.marker}${param.seriesName}: ${formattedVal}<br/>`;
-                    });
-                    return res;
-                }},
-                grid: { left: '3%', right: '4%', bottom: '3%', top: '15%', containLabel: true },
-                xAxis: { type: 'category', boundaryGap: false, data: timeData },
-                yAxis: { type: 'value', name: unit },
-                series: series
-            };
-            chartRequests.setOption(option);
-        }
-
-        //5.有使用
-        function updatePerformanceSection(results) {
-            const metrics = metricsConfig.performance;
-            const series = [];
-            let timeData = [];
-
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (!data || data.type !== 'time') return;
-                
-                // Update KPI (Avg)
-                document.getElementById(`kpi_${metric}`).innerText = data.avg.toFixed(2) + ' ms';
-
-                if (timeData.length === 0) timeData = data.timeData;
-
-                series.push({
-                    name: metricLabels[metric],
-                    type: 'line',
-                    smooth: true,
-                    data: data.valueData,
-                    itemStyle: { color: metricColors[metric] }
-                });
-            });
-
-            const option = {
-                tooltip: { trigger: 'axis', formatter: '{b}<br/>{a}: {c} ms' },
-                legend: { data: metrics.map(m => metricLabels[m]), bottom: 0 },
-                grid: { left: '3%', right: '4%', bottom: '10%', top: '15%', containLabel: true },
-                xAxis: { type: 'category', boundaryGap: false, data: timeData },
-                yAxis: { type: 'value', name: 'ms' },
-                series: series
-            };
-            chartPerformance.setOption(option);
-        }
-
-        function updateEdgeFunctionsSection(results) {
-            const metrics = metricsConfig.edgeFunctions;
-            
-            // Map metrics to specific charts and KPIs
-            const metricMap = {
-                'function_requestCount': {
-                    chart: chartFunctionRequests,
-                    kpiId: 'kpi_function_requestCount',
-                    unitType: 'count',
-                    color: metricColors['function_requestCount']
-                },
-                'function_cpuCostTime': {
-                    chart: chartFunctionCpu,
-                    kpiId: 'kpi_function_cpuCostTime',
-                    unitType: 'ms',
-                    color: metricColors['function_cpuCostTime']
-                }
-            };
-
-            metrics.forEach(metric => {
-                const config = metricMap[metric];
-                if (!config) return;
-
-                const data = results[metric];
-                if (!data || data.type !== 'time') return;
-
-                // Update KPI
-                const kpiEl = document.getElementById(config.kpiId);
-                if (kpiEl) {
-                    if (config.unitType === 'count') {
-                        kpiEl.innerText = formatCount(data.sum);
-                    } else if (config.unitType === 'ms') {
-                        // Format large numbers with commas
-                        kpiEl.innerText = data.sum.toLocaleString() + ' ms';
-                    }
-                }
-
-                // Chart
-                const chart = config.chart;
-                if (!chart) return;
-
-                let seriesData = data.valueData;
-                let unit = '';
-                let divisor = 1;
-
-                if (config.unitType === 'count') {
-                     const best = getBestCountUnit(Math.max(...data.valueData));
-                     unit = best.unit;
-                     divisor = best.divisor;
-                     seriesData = data.valueData.map(v => {
-                        const val = v / divisor;
-                        return (divisor === 1) ? val : val.toFixed(2);
-                     });
-                } else if (config.unitType === 'ms') {
-                     unit = 'ms';
-                }
-
-                const option = {
-                    tooltip: { trigger: 'axis', formatter: (params) => {
-                        let res = params[0].axisValue + '<br/>';
-                        params.forEach(param => {
-                            const val = data.valueData[param.dataIndex];
-                            let formattedVal = val;
-                            if (config.unitType === 'count') formattedVal = formatCount(val);
-                            else formattedVal = val.toLocaleString() + ' ms';
-                            
-                            res += `${param.marker}${param.seriesName}: ${formattedVal}<br/>`;
-                        });
-                        return res;
-                    }},
-                    legend: { data: [metricLabels[metric]], bottom: 0 },
-                    grid: { left: '3%', right: '4%', bottom: '10%', top: '15%', containLabel: true },
-                    xAxis: { type: 'category', boundaryGap: false, data: data.timeData },
-                    yAxis: { type: 'value', name: unit },
-                    series: [{
-                        name: metricLabels[metric],
-                        type: 'line',
-                        smooth: true,
-                        data: seriesData,
-                        itemStyle: { color: config.color },
-                        areaStyle: { opacity: 0.1 }
-                    }]
-                };
-                chart.setOption(option);
-            });
-        }
-
-        //5.1. Security Section
-        function updateSecuritySection(results) {
-            const metrics = metricsConfig.security;
-            
-            // Check time range limit
-            const { startTime, endTime } = calculateTimeRange();
-            const start = new Date(startTime);
-            const end = new Date(endTime);
-            // Allow a small buffer
-            if ((end - start) > (14 * 24 * 60 * 60 * 1000 + 60000)) {
-                const kpiEl = document.getElementById('kpi_security_hits');
-                if (kpiEl) {
-                    kpiEl.innerText = "范围过大";
-                    kpiEl.parentElement.querySelector('p').innerText = "仅支持查询14天内的数据";
-                    kpiEl.parentElement.querySelector('p').classList.add('text-red-500');
-                }
-                
-                chartSecurity.clear();
-                chartSecurity.setOption({
-                     title: {
-                         text: '该指标仅支持查询14天内的数据',
-                         left: 'center',
-                         top: 'center',
-                         textStyle: { color: '#9ca3af', fontSize: 14, fontWeight: 'normal' }
-                     }
-                });
-                return;
-            }
-            
-            // Reset error style if applicable
-            const kpiEl = document.getElementById('kpi_security_hits');
-            if (kpiEl) {
-                const p = kpiEl.parentElement.querySelector('p');
-                if (p) {
-                    p.innerText = "DDoS/CC 防护总拦截次数";
-                    p.classList.remove('text-red-500');
-                }
-            }
-
-            const series = [];
-            let timeData = [];
-            let totalHits = 0;
-
-            metrics.forEach(metric => {
-                const data = results[metric];
-                if (!data || data.type !== 'time') return;
-                
-                totalHits += data.sum;
-
-                if (timeData.length === 0) timeData = data.timeData;
-
-                series.push({
-                    name: metricLabels[metric],
-                    type: 'line',
-                    smooth: true,
-                    stack: 'Total', // Stacked area chart
-                    areaStyle: {},
-                    emphasis: { focus: 'series' },
-                    data: data.valueData,
-                    itemStyle: { color: metricColors[metric] }
-                });
-            });
-
-            // Update KPI
-            if (kpiEl) {
-                const formatted = formatCount(totalHits);
-                kpiEl.innerText = formatted;
-            }
-
-            const option = {
-                title: { show: false },
-                tooltip: { 
-                    trigger: 'axis', 
-                    axisPointer: { type: 'cross', label: { backgroundColor: '#6a7985' } }
-                },
-                legend: { data: metrics.map(m => metricLabels[m]), bottom: 0 },
-                grid: { left: '3%', right: '4%', bottom: '10%', top: '15%', containLabel: true },
-                xAxis: { type: 'category', boundaryGap: false, data: timeData },
-                yAxis: { type: 'value' },
-                series: series
-            };
-            chartSecurity.setOption(option);
-        }
-
-        //通用处理
-        const renderTopChart = (results, chartInstance, metricName, mapObject = null) => {
-            const data = results[metricName];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10);
-            
-            let unit = '';
-            let divisor = 1;
-            let label = metricLabels[metricName] || '';
-            
-            if (metricName.includes('outFlux')) {
-                const maxVal = sortedData.length > 0 ? sortedData[0].Value : 0;
-                const best = getBestUnit(maxVal, 'bytes');
-                unit = best.unit;
-                divisor = best.divisor;
-            } else {
-                const maxVal = sortedData.length > 0 ? sortedData[0].Value : 0;
-                const best = getBestCountUnit(maxVal);
-                unit = best.unit;
-                divisor = best.divisor;
-            }
-
-            const yAxisData = sortedData.map(item => mapObject?.[item.Key] || item.Key).reverse(); //问题点：没对“字段不存在”的短横杠做判断
-            const seriesData = sortedData.map(item => {
-                const val = item.Value / divisor;
-                return (divisor === 1) ? val : val.toFixed(2);
-            }).reverse();
-
-            let color = '#3b82f6';
-            const lowerName = metricName.toLowerCase();
-            if (lowerName.includes('country')) color = '#3b82f6';
-            else if (lowerName.includes('province') || lowerName.includes('resourcetype') || lowerName.includes('browser')) color = '#f59e0b';
-            else if (lowerName.includes('statuscode') || lowerName.includes('referer') || lowerName.endsWith('_ua')) color = '#8b5cf6';
-            else if (lowerName.includes('domain') || lowerName.includes('device')) color = '#06b6d4';
-            else if (lowerName.includes('url') || lowerName.includes('os')) color = '#10b981';
-            else if (lowerName.includes('sip')) color = '#ef4444';
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         let name = param.name;
-                         if (name.length > 100) name = name.substring(0, 100) + '...';
-                         
-                         const rawVal = sortedData[sortedData.length - 1 - param.dataIndex].Value;
-                         
-                         let formattedVal = '';
-                         if (metricName.includes('outFlux')) {
-                             formattedVal = formatBytes(rawVal);
-                         } else {
-                             const fVal = formatCount(rawVal);
-                             // Append '次' if not present in unit (formatCount returns unit, but we might want explicit '次')
-                             // formatCount returns "1.23 万". "1.23 万次" sounds good.
-                             // "500" -> "500 次".
-                             formattedVal = fVal + (fVal.includes('千') || fVal.includes('万') || fVal.includes('亿') ? '次' : ' 次');
-                         }
-                         
-                         return `${name}<br/>${param.marker}${label}: ${formattedVal}`;
-                    }
-                },
-                grid: { left: '3%', right: '10%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: unit },
-                yAxis: { 
-                    type: 'category', 
-                    data: yAxisData,
-                    axisLabel: {
-                        formatter: function (value) {
-                            if (value.length > 20) return value.substring(0, 20) + '...';
-                            return value;
-                        }
-                    }
-                },
-                series: [
-                    {
-                        name: label,
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: color },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-            chartInstance.setOption(option);
-        };
-
-        function updateTopMapChart(results) {
-            const metric = 'l7Flow_request_country';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Map data to ECharts format
-            // IMPORTANT: The 'name' here must match the region name AFTER nameMap is applied.
-            // Since worldNameMap maps English names to Chinese (e.g. "China" -> "中国大陆"),
-            // our data points must also use the Chinese names (e.g. "中国大陆") to match the map regions.
-            const mapData = data.data.map(item => {
-                const name = countryMap[item.Key] || codeToMapName[item.Key] || item.Key;
-                return {
-                    name: name,
-                    value: item.Value
-                };
-            });
-
-            const maxVal = mapData.length > 0 ? Math.max(...mapData.map(item => item.value)) : 0;
-
-            const option = {
-                tooltip: {
-                    trigger: 'item',
-                    formatter: function (params) {
-                        const val = params.value;
-                        return `${params.name}<br/>请求数: ${val ? val.toLocaleString() : 0} 次`;
-                    }
-                },
-                visualMap: {
-                    min: 0,
-                    max: maxVal,
-                    left: 'left',
-                    top: 'bottom',
-                    text: ['High', 'Low'],
-                    calculable: true,
-                    inRange: {
-                        color: ['#e0f2fe', '#0284c7']
-                    }
-                },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'map',
-                        mapType: 'world',
-                        roam: true,
-                        nameMap: worldNameMap,
-                        itemStyle: {
-                            emphasis: { label: { show: true } }
-                        },
-                        data: mapData
-                    }
-                ]
-            };
-
-            chartTopMap.setOption(option);
-        }
-
-        //6.有使用
-        function updateTopAnalysisSection(results) {
-            updateTopMapChart(results);
-
-            // Flux Charts
-            renderTopChart(results, chartTopCountry, 'l7Flow_outFlux_country', countryMap);
-            renderTopChart(results, chartTopProvince, 'l7Flow_outFlux_province', provinceMap);
-            renderTopChart(results, chartTopStatusCode, 'l7Flow_outFlux_statusCode');
-            renderTopChart(results, chartTopDomain, 'l7Flow_outFlux_domain');
-            renderTopChart(results, chartTopUrl, 'l7Flow_outFlux_url');
-            renderTopChart(results, chartTopResourceType, 'l7Flow_outFlux_resourceType');
-            renderTopChart(results, chartTopSip, 'l7Flow_outFlux_sip');
-            //renderTopChart(results, chartTopReferer, 'l7Flow_outFlux_referers'); //以前方案
-            updateTopRefererChart(results); //降级处理，包含“字段不存在”处理
-            renderTopChart(results, chartTopUaDevice, 'l7Flow_outFlux_ua_device');
-            renderTopChart(results, chartTopUaBrowser, 'l7Flow_outFlux_ua_browser');
-            renderTopChart(results, chartTopUaOs, 'l7Flow_outFlux_ua_os');
-            renderTopChart(results, chartTopUa, 'l7Flow_outFlux_ua');
-
-            // Request Charts
-            renderTopChart(results, chartTopRequestCountry, 'l7Flow_request_country', countryMap);
-            renderTopChart(results, chartTopRequestProvince, 'l7Flow_request_province', provinceMap);
-            renderTopChart(results, chartTopRequestStatusCode, 'l7Flow_request_statusCode');
-            renderTopChart(results, chartTopRequestDomain, 'l7Flow_request_domain');
-            renderTopChart(results, chartTopRequestUrl, 'l7Flow_request_url');
-            renderTopChart(results, chartTopRequestResourceType, 'l7Flow_request_resourceType');
-            renderTopChart(results, chartTopRequestSip, 'l7Flow_request_sip');
-            //renderTopChart(results, chartTopRequestReferer, 'l7Flow_request_referers');
-            updateTopRequestRefererChart(results); //降级方案
-            renderTopChart(results, chartTopRequestUaDevice, 'l7Flow_request_ua_device');
-            renderTopChart(results, chartTopRequestUaBrowser, 'l7Flow_request_ua_browser');
-            renderTopChart(results, chartTopRequestUaOs, 'l7Flow_request_ua_os');
-            renderTopChart(results, chartTopRequestUa, 'l7Flow_request_ua');
-        }
-
-
-
-        /**  
-         *  下面这一堆我也不知道是干嘛的
-         *  可能是二叉脆脆遗留下的降级方案吧
-         *  先不删了，省的再出问题
-         *  不过倒是可以先留着也是，说不定后面可能会用到？
-         *  真的太神秘了
-         * **/
-
-        function updateTopCountryChart(results) {
-            const metric = 'l7Flow_outFlux_country';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => countryMap[item.Key] || item.Key).reverse(); // Countries
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#3b82f6' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopCountry.setOption(option);
-        }
-
-        function updateTopProvinceChart(results) {
-            const metric = 'l7Flow_outFlux_province';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => provinceMap[item.Key] || item.Key).reverse(); // Map ID to Province Name
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#f59e0b' }, // Amber
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopProvince.setOption(option);
-        }
-
-        function updateTopStatusCodeChart(results) {
-            const metric = 'l7Flow_outFlux_statusCode';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse(); // Status Codes
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#8b5cf6' }, // Purple
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopStatusCode.setOption(option);
-        }
-
-        function updateTopDomainChart(results) {
-            const metric = 'l7Flow_outFlux_domain';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse(); // Domains
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#06b6d4' }, // Cyan
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopDomain.setOption(option);
-        }
-
-        function updateTopUrlChart(results) {
-            const metric = 'l7Flow_outFlux_url';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse(); // URLs
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         // Truncate long URLs in tooltip if needed, but usually full URL is better
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { 
-                    type: 'category', 
-                    data: yAxisData,
-                    axisLabel: {
-                        formatter: function (value) {
-                            // Truncate long URLs in axis label
-                            if (value.length > 30) {
-                                return value.substring(0, 30) + '...';
-                            }
-                            return value;
-                        }
-                    }
-                },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#10b981' }, // Green
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopUrl.setOption(option);
-        }
-
-        function updateTopResourceTypeChart(results) {
-            const metric = 'l7Flow_outFlux_resourceType';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse(); // Resource Types
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#f59e0b' }, // Amber
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopResourceType.setOption(option);
-        }
-
-        function updateTopSipChart(results) {
-            const metric = 'l7Flow_outFlux_sip';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse(); // Client IPs
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#ef4444' }, // Red
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopSip.setOption(option);
-        }
-
-        //无调用，作为备用方案来处理
-        function updateTopRefererChart(results) {
-            const metric = 'l7Flow_outFlux_referers';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => {
-                // Clean up key: remove backticks and trim
-                let key = item.Key.replace(/`/g, '').trim();
-                if (!key || key === '-') return '字段不存在'; //无/直接访问
-                return key;
-            }).reverse();
-            
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { 
-                    type: 'category', 
-                    data: yAxisData,
-                    axisLabel: {
-                        formatter: function (value) {
-                            // Truncate long URLs
-                            if (value.length > 30) {
-                                return value.substring(0, 30) + '...';
-                            }
-                            return value;
-                        }
-                    }
-                },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#8b5cf6' }, // Purple
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopReferer.setOption(option);
-        }
-
-        function updateTopUaDeviceChart(results) {
-            const metric = 'l7Flow_outFlux_ua_device';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#06b6d4' }, // Cyan
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopUaDevice.setOption(option);
-        }
-
-        function updateTopUaBrowserChart(results) {
-            const metric = 'l7Flow_outFlux_ua_browser';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#f59e0b' }, // Amber
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopUaBrowser.setOption(option);
-        }
-
-        function updateTopUaOsChart(results) {
-            const metric = 'l7Flow_outFlux_ua_os';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#10b981' }, // Green
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopUaOs.setOption(option);
-        }
-
-        function updateTopUaChart(results) {
-            const metric = 'l7Flow_outFlux_ua';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key.substring(0, 50) + (item.Key.length > 50 ? '...' : '')).reverse(); // Truncate long UAs
-            const seriesData = sortedData.map(item => (item.Value / (1024 * 1024 * 1024)).toFixed(2)).reverse(); // GB
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         // Find full UA string
-                         const originalItem = sortedData.find((item, index) => index === (sortedData.length - 1 - param.dataIndex));
-                         const fullName = originalItem ? originalItem.Key : param.name;
-                         return `${fullName}<br/>${param.marker}流量: ${param.value} GB`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: 'GB' },
-                yAxis: { type: 'category', data: yAxisData, axisLabel: { interval: 0, width: 200, overflow: 'truncate' } },
-                series: [
-                    {
-                        name: '流量',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#8b5cf6' }, // Purple
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopUa.setOption(option);
-        }
-
-        function updateTopRequestCountryChart(results) {
-            const metric = 'l7Flow_request_country';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => countryMap[item.Key] || item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#3b82f6' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestCountry.setOption(option);
-        }
-
-        function updateTopRequestProvinceChart(results) {
-            const metric = 'l7Flow_request_province';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => provinceMap[item.Key] || item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#f59e0b' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestProvince.setOption(option);
-        }
-
-        function updateTopRequestStatusCodeChart(results) {
-            const metric = 'l7Flow_request_statusCode';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#8b5cf6' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestStatusCode.setOption(option);
-        }
-
-        function updateTopRequestDomainChart(results) {
-            const metric = 'l7Flow_request_domain';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData, axisLabel: { interval: 0, width: 200, overflow: 'truncate' } },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#06b6d4' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestDomain.setOption(option);
-        }
-
-        function updateTopRequestUrlChart(results) {
-            const metric = 'l7Flow_request_url';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData, axisLabel: { interval: 0, width: 200, overflow: 'truncate' } },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#10b981' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestUrl.setOption(option);
-        }
-
-        function updateTopRequestResourceTypeChart(results) {
-            const metric = 'l7Flow_request_resourceType';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#f59e0b' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestResourceType.setOption(option);
-        }
-
-        function updateTopRequestSipChart(results) {
-            const metric = 'l7Flow_request_sip';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#8b5cf6' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestSip.setOption(option);
-        }
-
-        //无调用，作为备份使用
-        function updateTopRequestRefererChart(results) {
-            const metric = 'l7Flow_request_referers';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => {
-                 let key = item.Key;
-                 if (key === '-') return '字段不存在'; //直接访问
-                 // Clean up backticks and extra spaces
-                 key = key.replace(/`/g, '').trim();
-                 return key.substring(0, 50) + (key.length > 50 ? '...' : '');
-            }).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         // Find full referer string
-                         const originalItem = sortedData.find((item, index) => index === (sortedData.length - 1 - param.dataIndex));
-                         let fullName = param.name;
-                         if (originalItem) {
-                             let key = originalItem.Key;
-                             if (key === '-') fullName = '字段不存在'; //直接访问
-                             else fullName = key.replace(/`/g, '').trim();
-                         }
-                         return `${fullName}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData, axisLabel: { interval: 0, width: 200, overflow: 'truncate' } },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#ec4899' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestReferer.setOption(option);
-        }
-
-        function updateTopRequestUaDeviceChart(results) {
-            const metric = 'l7Flow_request_ua_device';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#06b6d4' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestUaDevice.setOption(option);
-        }
-
-        function updateTopRequestUaBrowserChart(results) {
-            const metric = 'l7Flow_request_ua_browser';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#10b981' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestUaBrowser.setOption(option);
-        }
-
-        function updateTopRequestUaOsChart(results) {
-            const metric = 'l7Flow_request_ua_os';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key).reverse();
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         return `${param.name}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#f59e0b' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestUaOs.setOption(option);
-        }
-
-        function updateTopRequestUaChart(results) {
-            const metric = 'l7Flow_request_ua';
-            const data = results[metric];
-
-            if (!data || data.type !== 'top' || !data.data) return;
-
-            // Sort by value (descending)
-            const sortedData = [...data.data].sort((a, b) => b.Value - a.Value).slice(0, 10); // Top 10
-
-            const yAxisData = sortedData.map(item => item.Key.substring(0, 50) + (item.Key.length > 50 ? '...' : '')).reverse(); // Truncate long UAs
-            const seriesData = sortedData.map(item => item.Value).reverse();
-
-            const option = {
-                tooltip: {
-                    trigger: 'axis',
-                    axisPointer: { type: 'shadow' },
-                    formatter: (params) => {
-                         const param = params[0];
-                         // Find full UA string
-                         const originalItem = sortedData.find((item, index) => index === (sortedData.length - 1 - param.dataIndex));
-                         const fullName = originalItem ? originalItem.Key : param.name;
-                         return `${fullName}<br/>${param.marker}请求数: ${param.value.toLocaleString()} 次`;
-                    }
-                },
-                grid: { left: '3%', right: '4%', bottom: '3%', containLabel: true },
-                xAxis: { type: 'value', name: '次' },
-                yAxis: { type: 'category', data: yAxisData, axisLabel: { interval: 0, width: 200, overflow: 'truncate' } },
-                series: [
-                    {
-                        name: '请求数',
-                        type: 'bar',
-                        data: seriesData,
-                        itemStyle: { color: '#8b5cf6' },
-                        label: { show: true, position: 'right' }
-                    }
-                ]
-            };
-
-            chartTopRequestUa.setOption(option);
-        }
-        /** 神秘降级方案结束 **/
-
-        // Handle Resize
-        window.addEventListener('resize', () => {
-            chartTraffic && chartTraffic.resize();
-            chartBandwidth && chartBandwidth.resize();
-            chartRequests && chartRequests.resize();
-            chartPerformance && chartPerformance.resize();
-            chartOriginPull && chartOriginPull.resize();
-            chartTopCountry && chartTopCountry.resize();
-            chartTopProvince && chartTopProvince.resize();
-            chartTopStatusCode && chartTopStatusCode.resize();
-            chartTopDomain && chartTopDomain.resize();
-            chartTopUrl && chartTopUrl.resize();
-            chartTopResourceType && chartTopResourceType.resize();
-            chartTopSip && chartTopSip.resize();
-            chartTopReferer && chartTopReferer.resize();
-            chartTopUaDevice && chartTopUaDevice.resize();
-            chartTopUaBrowser && chartTopUaBrowser.resize();
-            chartTopUaOs && chartTopUaOs.resize();
-            chartTopUa && chartTopUa.resize();
-            chartTopRequestCountry && chartTopRequestCountry.resize();
-            chartTopRequestProvince && chartTopRequestProvince.resize();
-            chartTopRequestStatusCode && chartTopRequestStatusCode.resize();
-            chartTopRequestDomain && chartTopRequestDomain.resize();
-            chartTopRequestUrl && chartTopRequestUrl.resize();
-            chartTopRequestResourceType && chartTopRequestResourceType.resize();
-            chartTopRequestSip && chartTopRequestSip.resize();
-            chartTopRequestReferer && chartTopRequestReferer.resize();
-            chartTopRequestUaDevice && chartTopRequestUaDevice.resize();
-            chartTopRequestUaBrowser && chartTopRequestUaBrowser.resize();
-            chartTopRequestUaOs && chartTopRequestUaOs.resize();
-            chartTopRequestUa && chartTopRequestUa.resize();
-            chartTopMap && chartTopMap.resize();
-        });
-
-        // Start - Entrance
-        initDashboard();
-    
+import './dashboard/main.js';
