+++ index.html
@@ -725,6 +725,6 @@
 
     </div>
 
-    <script src="./assets/dashboard.js"></script>
+    <script type="module" src="./assets/dashboard.js"></script>
 </body>
 </html>
